create separate search component
